functions/care-symbols-infer/requirements.txt
appwrite==6.1.0
ai-edge-litert==1.0.1
numpy==1.26.4
Pillow==10.4.0
requests==2.31.0